Redirect to the new website
Eventually, this page should be taken down. But folks might have
it bookmarked, and other websites link to it. Therefore, taking
it down could cause confusion if the link is broken.

Instead, this adds a redirect so that visitors to this page will
be taken to the new site.

In a few months (say, 3-6 months), it will probably be safe to
completely take down this page.